requirements.txt
perplexityai==0.13.0
structlog
requests
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0
langchain-community>=0.1.0
pydantic>=2.0.0
typing-extensions>=4.0.0
dataclasses-json>=0.6.0
python-dotenv>=1.0.0
fastapi>=0.100.0
google-genai
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
azure-functions>=1.18.0
google-search-results==2.4.2
# Azure specific dependencies (for Key Vault and monitoring)
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0
azure-monitor-opentelemetry>=1.0.0